Store slackbot token in browser localStorage & drop crypto-js
We have to give the plain-text token to the server anyway so it can
present that token to slack's API. So server side was going to either
have the cyphertext & the encryption key both on disk, or we can just
not save that on disk at all.

Browser localStorage is not ideal. But it is another case where the
cyphertext and the encryption key must both be accessible to the code
anyway, so localStorage is only marginally less secure.

Also simplified parsing of html form by using FormData object instead
of reducing the form.elements myself.